[feature] #2547: Add logo to `AssetDefinition`
